+++ packages/plugin-page-view-tracking-browser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.45](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.3.44...@amplitude/plugin-page-view-tracking-browser@2.3.45) (2025-08-28)
+
+**Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser
+
+
+
+
+
+## [2.3.44](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.3.43...@amplitude/plugin-page-view-tracking-browser@2.3.44) (2025-08-26)
+
+**Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser
+
+
+
+
+
 ## [2.3.43](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.3.42...@amplitude/plugin-page-view-tracking-browser@2.3.43) (2025-08-25)
 
 **Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser